fix: replace svg-url-loader with @svgr/webpack
Danni had tried to replace react-svg-loader with svg-url-loader, but the two are not interchangeable at all and do different things. I just used what we use in core, @svgr/webpack, with the same config as core too. This works well and no longer breaks the addon!

Also bumped the Local core dependency as well.