fix: only copy files we want from the tmp dir
This handles old backups that may still have the log file included. I just used the same glob helper to get the list of files we want to inlcude from the backup and made sure to only copy those. This finally fixes things on windows without stopping the site in the meantime.